Add a pool containing unprocessed deferred destroy operations.
The pool contains the "bptree_obj" entry in the MOS object directory.
It's DMU object contains a pair of bptree_entry_phys_t items, one for
each of a pair of destroyed small filesystems.  Although ZoL 0.6.2 was
used to create the pool, it was modified to not process the deferred
destroy list in order to cleanly produce a pool in this state.