Add minimal reproducer for ZoL issue 4083
Adds a test pool which designed to be a minimal reproducer of
https://github.com/zfsonlinux/zfs/issues/4083.  The pool contains
a single filesystem "tank" with only a file "file2" in it. When we try
to delete the file we trigger an assertion in zfs_remove(): this is
because the file xattr obj was already deleted. This data discrepancy
is not detected by a scrub which completes successfully.

Signed-off-by: loli10K <ezomori.nozomu@gmail.com>
Reviewed-by: Brian Behlendorf <behlendorf1@llnl.gov>
Closes #3